Use GIT_* envvars for author and committer
Something is overwriting the effect of the git config calls in
.travis.yml. Tests that need a committer and/or author are failing with

    empty ident name (for <...@...>) not allowed

Set GIT_COMMITTER_NAME, GIT_COMMITTER_EMAIL, GIT_AUTHOR_NAME, and
GIT_AUTHOR_EMAIL so no ident name will be empty. Besides fixing the
failures, this isolates the effect to the test process, instead of
depending on an external config file.