Call higher level merge_base() to compare branches
The previous comparison of the merge base to either of the branches was not
correct.

- git.repo.Repo.rev_parse() returns a Commit object (in this case)

- self.repo.git is of type git.cmd.Git, and Git.merge_base() returns a unicode
  object.

- Comparing these two returns False even in cases in which the intended
  comparision should be True. This causes _compare_branches() to return 3
  (Branch needs a real merge) when it ought to be returning either 1 (First
  given branch needs fast-forwarding) or 2 (Second given branch needs
  fast-forwarding)

- A consequence of this is that a call to require_branches_equal() (e.g. by
  must_be_uptodate(), during finish()) exits with a failure claiming that a
  merge is needed, even in a case (status == 2) in which a difference would be
  allowed (the local branch being ahead of the origin).

Substitute the call to self.repo.git.merge_base() with a call to
self.git.merge_base() to invoke the function git.repo.Repo.merge_base() that
is a wrapper around the lower level merge_base().

This is almost, but not quite, sufficient. It returns a list of Commit
objects, which will have at max one Commit. Access that Commit at index [0].

The list will be empty if no common merge base exists. Handle what ought to be
an uncommon case by adding IndexError to the appropriate case (return value
4).

An alternative solution would have been to keep the original lower level
function call that returned a unicode object for base, but use commit1.hexsha
and commit2.hexsha when comparing it to the Commit objects corresponding to
branch1 and branch2 respectively. Using a type with more information for the
comparison is preferred.