Skip create tag during finish if it already exists
The tagging during finish had a comment copied directly copied from the source
of the bash implementation of git-flow:

	In case a previous attempt to finish this release branch
   has failed, but the merge into develop was successful, we skip it
   now.

But this had not been implemented. As long as tagging_info was set, the tag
creation was unconditional. This change conditions it on the tag not yet
existing.

Without this change, the call to finish is not idempotent. Consider this
sequence:
* User calls finish
* Finish tags the repo
* Finish fails merging back to develop due to merge conflict
* User resolves merge conflict
* User calls finish again
* Finish fails tagging the repo because the tag already exists

By conditioning the tag, finish is idempotent. The second tag attempt is a
NOP.

Another commented acknowledged the non-implementation (via ":todo:") of a
check that if the tag exists, it points to the commit. This change implements
this check as well. Raise TagExistsError if the check fails.